Add leadership, committee, and person schemas
