+++ schemas/objects/person.ts
@@ -1,27 +1,15 @@
-import { DocumentDefinition } from "sanity"
+import { ObjectDefinition } from "sanity"
 
-const schema: DocumentDefinition = {
+const schema: ObjectDefinition = {
     name: "person",
     title: "Person",
-    type: "document",
+    type: "object",
     fields: [
         {
             name: "name",
             title: "Name",
             type: "string",
             validation: (Rule) => Rule.required(),
-        },
-        {
-            name: "boardYear",
-            title: "Board Year",
-            type: "reference",
-            to: [{ type: "campYear" }],
-            validation: (Rule) => Rule.required(),
-        },
-        {
-            name: "committee",
-            title: "Committee/Org Level",
-            type: "string",
         },
         {
             name: "position",
@@ -32,18 +20,15 @@ const schema: DocumentDefinition = {
             name: "propic",
             title: "Profile Picture",
             type: "image",
+            options: {
+              hotspot: true,
+            },
         },
         {
             // see https://www.sanity.io/docs/block-type
             name: "description",
             title: "Description",
             type: "portableText",
-        },
-        {
-            name: "order",
-            title: "Order",
-            type: "number",
-            hidden: true,
         },
     ],
 }
